organizing code for esteemer
